# server/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Monitoring Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "montoring.nought.ru",
    "basePath": "/api/v1",
    "paths": {
        "/agents": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Get all agents",
                "operationId": "getAllAgents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Agent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Create new agent",
                "operationId": "createAgent",
                "parameters": [
                    {
                        "description": "Create agent body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateAgentBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CreateAgentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/agents/{agentID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Get agent by ID",
                "operationId": "getAgentByID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Agent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/agents/{agentID}/frames": {
            "get": {
                "produces": [
                    "multipart/x-mixed-replace"
                ],
                "tags": [
                    "streaming"
                ],
                "summary": "Get stream frames",
                "operationId": "getStreamFrames",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "MJPEG stream",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/agents/{agentID}/setupconfig": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Download agent files",
                "operationId": "downloadAgentFiles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Download agent files",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AgentConfigBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/agents/{agentID}/specifications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "Get agent specifications",
                "operationId": "getAgentSpecs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent ID",
                        "name": "agentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AgentSpecs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Agent": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isOnline": {
                    "type": "boolean"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "AgentConfigBody": {
            "type": "object",
            "properties": {
                "enrollmentKey": {
                    "type": "string"
                }
            }
        },
...
        "AgentSpecs": {
            "type": "object",
            "properties": {
                "agentId": {
                    "type": "string"
                },
                "cpuSpecs": {
                    "$ref": "#/definitions/CpuSpecs"
                },
                "diskSpecs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DiskSpecs"
                    }
                },
                "hostSpecs": {
                    "$ref": "#/definitions/HostSpecs"
                },
                "memorySpecs": {
                    "$ref": "#/definitions/MemorySpecs"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "CpuSpecs": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "availability": {
                    "type": "string"
                },
                "currentClockSpeed": {
                    "type": "integer"
                },
                "data_width": {
                    "type": "integer"
                },
                "l2CacheSize": {
                    "type": "integer"
                },
                "l3CacheSize": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "maxClockSpeed": {
                    "type": "integer"
                },
                "modelName": {
                    "type": "string"
                },
                "numberOfCores": {
                    "type": "integer"
                },
                "numberOfEnabledCores": {
                    "type": "integer"
                },
                "numberOfLogicalProcessors": {
                    "type": "integer"
                },
                "processorId": {
                    "type": "string"
                },
                "socketDesignation": {
                    "type": "string"
                },
                "stepping": {
                    "type": "string"
                },
                "virtualizationFirmwareEnabled": {
                    "type": "boolean"
                }
            }
        },
        "CreateAgentBody": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "CreateAgentResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enrollmentKey": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isOnline": {
                    "type": "boolean"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "DiskSpecs": {
            "type": "object",
            "properties": {
                "device": {
                    "type": "string"
                },
                "fsType": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "HostSpecs": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string"
                },
                "kernelVersion": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "osArch": {
                    "type": "string"
                },
                "osType": {
                    "type": "string"
                },
                "osVersion": {
                    "type": "string"
                }
            }
        },
        "MemorySpecs": {
            "type": "object",
            "properties": {
                "physicalMemoryInfo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PhysicalMemoryInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "PhysicalMemoryInfo": {
            "type": "object",
            "properties": {
                "bankLabel": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "configuredClockSpeed": {
                    "type": "integer"
                },
                "deviceLocator": {
                    "type": "string"
                },
                "form_factor": {
                    "type": "string"
                },
                "hotSwappable": {
                    "type": "boolean"
                },
                "manufacturer": {
                    "type": "string"
                },
                "memoryType": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string"
                },
                "removable": {
                    "type": "boolean"
                },
                "replaceable": {
                    "type": "boolean"
                },
                "serialNumber": {
                    "type": "string"
                },
                "speed": {
                    "type": "integer"
                }
            }
        },
...
    }
}